pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "logux"
version = "1.0.0"
description = "Android Logs & Traffic CLI — real-time logcat viewer with filtering, presets, traffic inspection, and mock rules"
requires-python = ">=3.10"
license = "MIT"
authors = [
    { name = "ThothLab" },
]

dependencies = [
    "prompt-toolkit>=3.0",
    "rich>=13.0",
    "pyyaml>=6.0",
    "mitmproxy>=10.0",
    "watchdog>=3.0",
]

[project.scripts]
logux = "logux.__main__:main"

[tool.hatch.build.targets.wheel]
packages = ["logux"]
